refactor files and renamed them for better dx
