+++ src/lib/page-components/oyster/operator/OysterOperatorJobsHistory.svelte
@@ -1,37 +1,27 @@
 <script lang="ts">
-	import Pagination from '$lib/components/pagination/Pagination.svelte';
 	import SearchBar from '$lib/components/search/SearchBar.svelte';
+
+	import Pagination from '$lib/components/pagination/Pagination.svelte';
 	import PageTitle from '$lib/components/texts/PageTitle.svelte';
 	import { oysterStore } from '$lib/data-stores/oysterStore';
-	import { connected } from '$lib/data-stores/walletProviderStore';
 	import type {
 		OysterInventoryDataModel,
-		OysterInventorySortKeys
+		OysterOperatorInventorySortKeys
 	} from '$lib/types/oysterComponentType';
-	import { OYSTER_HISTORY_TABLE_HEADER } from '$lib/utils/constants/oysterConstants';
-	import { getSearchedInventoryData, sortOysterInventory } from '$lib/utils/helpers/oysterHelpers';
+	import { OYSTER_OPERATOR_HISTORY_TABLE_HEADER } from '$lib/utils/constants/oysterConstants';
+	import {
+		getSearchedOysterJobsData,
+		sortOysterOperatorHistory
+	} from '$lib/utils/helpers/oysterHelpers';
 	import OysterTableCommon from '$lib/page-components/oyster/inventory/OysterTableCommon.svelte';
-	import OysterHistoryTableRow from '$lib/page-components/oyster/history/OysterHistoryTableRow.svelte';
+	import OysterOperatorJobsHistoryTableRow from '$lib/page-components/oyster/operator/OysterOperatorJobsHistoryTableRow.svelte';
+	import { OYSTER_OPERATOR_JOBS_URL } from '$lib/utils/constants/urls';
 	import { TABLE_ITEMS_PER_PAGE } from '$lib/utils/constants/constants';
-	import { OYSTER_OWNER_INVENTORY_URL } from '$lib/utils/constants/urls';
 
 	let searchInput = '';
 	let activePage = 1;
+	let operatorHistoryData: OysterInventoryDataModel[] | undefined;
 	let sortingMap: Record<string, 'asc' | 'desc'> = {};
-	let inventoryData: OysterInventoryDataModel[] | undefined;
-
-	const handleSortData = (id: string) => {
-		if (sortingMap[id]) {
-			sortingMap[id] = sortingMap[id] === 'asc' ? 'desc' : 'asc';
-		} else {
-			sortingMap[id] = 'asc';
-		}
-		inventoryData = sortOysterInventory(
-			inventoryData,
-			id as OysterInventorySortKeys,
-			sortingMap[id]
-		);
-	};
 
 	const handlePageChange = (page: number) => {
 		activePage = page;
@@ -41,12 +31,24 @@
 		activePage = 1;
 	};
 
-	// fiter inventory data based on job status
-	$: inventoryData = $oysterStore.jobsData?.filter(
-		(data) => data.status === 'completed' || data.status === 'closed'
+	const handleSortData = (id: string) => {
+		if (sortingMap[id]) {
+			sortingMap[id] = sortingMap[id] === 'asc' ? 'desc' : 'asc';
+		} else {
+			sortingMap[id] = 'asc';
+		}
+		operatorHistoryData = sortOysterOperatorHistory(
+			operatorHistoryData,
+			id as OysterOperatorInventorySortKeys,
+			sortingMap[id]
+		);
+	};
+
+	$: operatorHistoryData = $oysterStore.merchantJobsData?.filter(
+		(data) => data.status === 'closed'
 	);
 	// get searched data based on searchInput
-	$: searchedData = getSearchedInventoryData(searchInput, inventoryData);
+	$: searchedData = getSearchedOysterJobsData(searchInput, operatorHistoryData);
 	// get page array based on inventory and itemsPerPage
 	$: pageCount = Math.ceil((searchedData?.length ?? 0) / TABLE_ITEMS_PER_PAGE);
 	// get paginated data based on activePage
@@ -57,31 +59,30 @@
 </script>
 
 <div class="mx-auto">
-	<PageTitle title="My Past Orders" backHref={OYSTER_OWNER_INVENTORY_URL} />
+	<PageTitle title="History of Claims" backHref={OYSTER_OPERATOR_JOBS_URL} />
 	<div class="mb-6 flex items-center gap-4">
 		<SearchBar
-			disabled={!$connected}
 			{onSearchClick}
 			bind:input={searchInput}
-			placeholder="Search for operator, instance or region"
+			placeholder="Search for user, instance or region"
 			styleClass="w-full"
 		/>
 	</div>
 	<OysterTableCommon
 		{handleSortData}
-		loading={!$oysterStore.oysterStoreLoaded}
-		tableHeading={OYSTER_HISTORY_TABLE_HEADER}
+		loading={!$oysterStore.merchantJobsLoaded}
+		tableHeading={OYSTER_OPERATOR_HISTORY_TABLE_HEADER}
 		noDataFound={paginatedData?.length ? false : true}
 	>
 		{#if paginatedData?.length}
 			{#each paginatedData as rowData, rowIndex}
-				<OysterHistoryTableRow {rowData} {rowIndex} />
+				<OysterOperatorJobsHistoryTableRow {rowData} {rowIndex} />
 			{/each}
+			<tr>
+				<td colspan="12">
+					<Pagination {pageCount} {activePage} {handlePageChange} styleClass="mt-6" />
+				</td>
+			</tr>
 		{/if}
-		<tr>
-			<td colspan="12">
-				<Pagination {pageCount} {activePage} {handlePageChange} styleClass="mt-6" />
-			</td>
-		</tr>
 	</OysterTableCommon>
 </div>
