+++ src/lib/page-components/oyster/operator/OysterOperatorJobsTableRow.svelte
