+++ src/lib/page-components/oyster/operator/OysterOperatorRegisterPage.svelte
