+++ src/routes/oyster/operator/jobs/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import OysterMerchantJobs from '$lib/page-components/oyster/operator/OysterOperatorJobs.svelte';
+	import OysterOperatorJobs from '$lib/page-components/oyster/operator/OysterOperatorJobs.svelte';
 </script>
 
-<OysterMerchantJobs />
+<OysterOperatorJobs />
